code refactoring and address the comment
