+++ src/promptflow-evals/promptflow/evals/evaluate/_batch_run_client/__init__.py
@@ -1,4 +1,5 @@
-from .code_client import BatchRunContext, CodeClient
+from .batch_run_context import BatchRunContext
+from .code_client import CodeClient
 from .proxy_client import ProxyClient
 
 __all__ = ["CodeClient", "ProxyClient", "BatchRunContext"]
