+++ src/promptflow-evals/promptflow/evals/evaluate/_batch_run_client/proxy_client.py
@@ -8,6 +8,8 @@ import numpy as np
 from promptflow.client import PFClient
 from promptflow.tracing import ThreadPoolExecutorWithContext as ThreadPoolExecutor
 
+from ..._constants import BATCH_RUN_TIMEOUT
+
 LOGGER = logging.getLogger(__name__)
 
 
@@ -28,7 +30,7 @@ class ProxyClient:
         return ProxyRun(run=eval_future)
 
     def get_details(self, proxy_run, all_results=False):
-        run = proxy_run.run.result(timeout=60 * 60)
+        run = proxy_run.run.result(timeout=BATCH_RUN_TIMEOUT)
         result_df = self._pf_client.get_details(run, all_results=all_results)
         result_df.replace("(Failed)", np.nan, inplace=True)
         return result_df
